requirements.txt
dash
dash-bootstrap-components
plotly
pandas
gunicorn
PyGithub
spacy
python-dotenv